mental_health_chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
# Database
sqlalchemy==2.0.23
# AI and NLP
transformers==4.35.2
torch==2.1.1
# Audio processing
openai-whisper==20231117
gtts==2.4.0
pydub==0.25.1
# Data visualization
matplotlib==3.8.2
seaborn==0.13.0
# Additional utilities
python-dateutil==2.8.2
numpy==1.24.3
pandas==2.1.4
